requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
crispy-bootstrap4==2022.1
cryptography==41.0.3
cssselect2==0.7.0
Django==4.2.4
django-crispy-forms==2.0
django-filter==23.2
django-quill-editor==0.1.40
djangorestframework==3.14.0
fonttools==4.42.0
gobject==0.1.0
html5lib==1.1
idna==3.4
lxml==4.9.3
oscrypto==1.3.0
Pillow==10.0.0
pycparser==2.21
pydyf==0.7.0
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pypdf==3.15.0
pyphen==0.14.0
pypng==0.20220715.0
python-barcode==0.15.1
python-bidi==0.4.2
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.4
weasyprint==59.0
webencodings==0.5.1
xhtml2pdf==0.2.11
zopfli==0.2.2
